+++ src/components/TodoList/TodoList.tsx
@@ -3,10 +3,13 @@ import styled from "styled-components";
 import AddTask from "../AddTask/AddTask";
 import TodoProvider from "src/context/TodoProvider";
 import Tasks from "../Tasks/Tasks";
+import SearchTasks from "../SearchTasks/SearchTasks";
 
-const TodoListStyle = styled.div`
+const TodoListWrapper = styled.div`
     display: flex;
-    flex-direction: column
+    flex-direction: column;
+    max-width: 800px;
+    margin: auto;
 `;
 
 const TasksTitle = styled.h3 `
@@ -15,15 +18,16 @@ const TasksTitle = styled.h3 `
 `
 
 const TodoList: React.FC = () => {
-    return <TodoListStyle className="todo-list">
+    return <TodoListWrapper className="todo-list">
         <TodoProvider>
             <TasksTitle>Tasks</TasksTitle>
 
+            <SearchTasks />
             <Tasks />
 
             <AddTask />
         </TodoProvider>
-    </TodoListStyle>
+    </TodoListWrapper>
 }
 
 export default TodoList;